+++ app.js
@@ -1,13 +1,10 @@
 require("dotenv").config({ path: `.env.${process.env.NODE_ENV}` });
 
 const base = require("./lib/base");
-
-const Transfer = require("./services/transfer");
+const Gsubz = require("./services");
 
 module.exports = (key) => {
   const request = base(key);
 
-  return {
-    Transfer: new Transfer(request),
-  };
+  return new Gsubz(request);
 };